requirements.txt
ccxt
pandas
numpy
scipy
pydantic
PyYAML
sqlalchemy
alembic
apscheduler
optuna
mlflow
streamlit
plotly
